rewriting th checkIfPointExists methode and modified Sensor related classes and interfaces.
	modified:   src/main/java/ohdm/storage/IdVerification.java
	modified:   src/main/java/ohdm/storage/IdVerificationInterface.java
	modified:   src/main/java/ohdm/storage/ImportedSensorDb.java
	modified:   src/main/java/ohdm/storage/ImportedSensorInterface.java
	modified:   src/main/java/ohdm/storage/PointsDb.java
	modified:   src/main/java/ohdm/storage/PointsInterface.java
	modified:   src/main/java/ohdm/storage/Sensor.java
	modified:   src/main/java/ohdm/storage/SensorInterface.java

+++ src/main/java/ohdm/storage/ImportedSensorDb.java
@@ -33,7 +33,7 @@ public class ImportedSensorDb {
         statement.close();
     }
 
-    public boolean checkIfIdExists(int importedSensorId) throws SQLException {
+    public boolean checkIfIdExists(long importedSensorId) throws SQLException {
         PreparedStatement statement = db.connection.prepareStatement("SELECT * FROM ohdm.imported_sensor "
                 + "WHERE imported_id = " + importedSensorId + ";");
         resultSet = statement.executeQuery();
